move Matrix3x3f to types.h, add direction to GpuVolume
